+++ CHANGELOG.md
@@ -11,7 +11,8 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
 ### Added
 
 - Add `color.ReportFlow` which is an extension of `middleware.ReportFlow`.
-- Add `otelgoyek.Instrument` which adds OpenTelemetry tracing instrumentation.
+- Add `otelgoyek.Middleware` and `otelgoyek.ExecutorMiddleware` which add
+  OpenTelemetry tracing instrumentation.
 
 ### Changed
 
@@ -22,7 +23,7 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
 
 - Drop support for Go 1.17, 1.18, 1.19, 1.20.
 
-## [0.1.7](https://github.com/goyek/x/releases/tag/v0.1.7) - 2024-01-17
+## [0.1.7](https://github.com/goyek/x\/releases/tag/v0.1.7) - 2024-01-17
 
 ### Added